# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zimbabwe Skill Verification Platform API
  version: 1.0.0
  description: API for institution-managed credentials, public verification, blockchain evidence, auditability, and analytics. Examples contain placeholders only. The uploaded PDF is authoritative; a generated PDF is a presentation copy. Credentials become active only after blockchain confirmation, and revoked credentials remain publicly verifiable.
  contact:
    name: API Support
    email: api-support@example.org
  license:
    name: ISC
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.example.org
    description: Production placeholder
...
paths:
...
  /api/credentials:
    get: { tags: [Credentials], summary: List institution-scoped credentials, operationId: listCredentials, security: [ { BearerAuth: [] } ], x-roles: [super_admin, institution_admin, issuer, verifier], parameters: [ { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Search' }, { $ref: '#/components/parameters/SortOrder' }, { name: sortBy, in: query, schema: { type: string, enum: [created_at, updated_at, issue_date, qualification, status] } }, { name: status, in: query, schema: { $ref: '#/components/schemas/CredentialStatus' } }, { name: institutionId, in: query, schema: { type: string, format: uuid } }, { name: studentId, in: query, schema: { type: string, format: uuid } }, { name: issueDateFrom, in: query, schema: { type: string, format: date } }, { name: issueDateTo, in: query, schema: { type: string, format: date } } ], responses: { '200': { $ref: '#/components/responses/Paginated' }, '403': { $ref: '#/components/responses/AccessDenied' } } }
...
  /api/credentials/issue:
    post:
      tags: [Credentials]
      summary: Issue a credential from an authoritative PDF
      description: Validates PDF bytes, hashes the original, uploads to IPFS, confirms the blockchain transaction, then activates the credential. Duplicate hashes return 409. Provider failures leave a controlled failed-processing status.
      operationId: issueCredential
      security: [ { BearerAuth: [] } ]
      x-roles: [super_admin, institution_admin, issuer]
      requestBody: { required: true, content: { multipart/form-data: { schema: { type: object, required: [certificate, studentId, institutionId, qualification, issueDate], properties: { certificate: { type: string, format: binary, description: PDF only; default maximum 10 MB }, studentId: { type: string, format: uuid }, institutionId: { type: string, format: uuid }, qualification: { type: string, minLength: 2, maxLength: 300 }, issueDate: { type: string, format: date } } } } } }
      responses: { '201': { $ref: '#/components/responses/Created' }, '400': { $ref: '#/components/responses/ValidationError' }, '403': { $ref: '#/components/responses/AccessDenied' }, '409': { $ref: '#/components/responses/Conflict' }, '413': { $ref: '#/components/responses/FileTooLarge' }, '422': { $ref: '#/components/responses/ValidationError' }, '503': { $ref: '#/components/responses/Unavailable' }, '504': { $ref: '#/components/responses/Unavailable' } }
  /api/credentials/{id}/revoke:
    patch: { tags: [Credentials], summary: Permanently revoke credential with blockchain confirmation, operationId: revokeCredential, security: [ { BearerAuth: [] } ], x-roles: [super_admin, institution_admin], parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { required: true, content: { application/json: { schema: { type: object, required: [reason], properties: { reason: { type: string, minLength: 5, maxLength: 1000 } }, additionalProperties: false } } } }, responses: { '200': { $ref: '#/components/responses/Success' }, '400': { $ref: '#/components/responses/ValidationError' }, '403': { $ref: '#/components/responses/AccessDenied' }, '409': { $ref: '#/components/responses/Conflict' }, '503': { $ref: '#/components/responses/Unavailable' } } }
  /api/credentials/{id}/generate-pdf:
    post: { tags: [Generated Certificates], summary: Generate a presentation certificate, description: The generated PDF is a presentation copy; the originally uploaded PDF remains authoritative., operationId: generateCredentialPdf, security: [ { BearerAuth: [] } ], x-roles: [super_admin, institution_admin, issuer], parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/AccessDenied' }, '409': { $ref: '#/components/responses/Conflict' } } }
  /api/credentials/{id}/pdf:
    get: { tags: [Generated Certificates], summary: Download the generated presentation PDF, description: Streams only the server-generated presentation copy as an attachment. The original uploaded PDF remains authoritative in IPFS and no server path is exposed., operationId: downloadCredentialPdf, security: [ { BearerAuth: [] } ], x-roles: [super_admin, institution_admin, issuer, verifier], parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { description: Presentation PDF attachment, headers: { Content-Disposition: { schema: { type: string }, description: Fixed safe attachment filename } }, content: { application/pdf: { schema: { type: string, format: binary } } } }, '403': { $ref: '#/components/responses/AccessDenied' }, '404': { $ref: '#/components/responses/NotFound' } } }
...
components:
  securitySchemes:
    BearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    CookieAuth: { type: apiKey, in: cookie, name: verifyzw_session, description: HttpOnly browser session cookie; cookie-authenticated unsafe requests also require X-CSRF-Token. }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Search: { name: search, in: query, schema: { type: string, maxLength: 200 } }
    SortOrder: { name: sortOrder, in: query, schema: { type: string, enum: [asc, desc], default: desc } }
...
  schemas:
...
    CredentialStatus: { type: string, enum: [pending, processing, active, failed, revoked] }
...
    StandardError: { type: object, required: [success, message], properties: { success: { type: boolean, enum: [false] }, code: { type: string, example: ACCESS_DENIED }, message: { type: string, example: You do not have permission to perform this action. }, requestId: { type: string, format: uuid, nullable: true } } }
    ValidationError: { allOf: [ { $ref: '#/components/schemas/StandardError' }, { type: object, properties: { code: { type: string, enum: [VALIDATION_ERROR] }, errors: { type: array, items: { type: object, required: [field, message], properties: { field: { type: string, example: email }, message: { type: string, example: A valid email address is required. } } } } } } ] }
    Pagination: { type: object, required: [page, limit, total, totalPages, hasNextPage, hasPreviousPage], properties: { page: { type: integer }, limit: { type: integer }, total: { type: integer }, totalPages: { type: integer }, hasNextPage: { type: boolean }, hasPreviousPage: { type: boolean } } }
...
  responses:
    Success: { description: Successful operation, content: { application/json: { schema: { type: object, required: [success], properties: { success: { type: boolean, enum: [true] } }, additionalProperties: true } } } }
    Created: { description: Resource created, content: { application/json: { schema: { type: object, required: [success], properties: { success: { type: boolean, enum: [true] } }, additionalProperties: true } } } }
    Paginated: { description: Paginated result; non-super-admin results are institution scoped where applicable, content: { application/json: { schema: { type: object, required: [success, pagination], properties: { success: { type: boolean, enum: [true] }, pagination: { $ref: '#/components/schemas/Pagination' } }, additionalProperties: true } } } }
...
    ValidationError: { description: Request validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ValidationError' } } } }
...
    AccessDenied: { description: Authenticated actor lacks permission or institution scope, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
...
    NotFound: { description: Resource not found or feature disabled, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
    Conflict: { description: Duplicate or conflicting resource state, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
...
    FileTooLarge: { description: Configured upload size limit exceeded, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
...
    Unavailable: { description: Dependency unavailable or operation timed out, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }